Split the main reader into static/dynamic cases
The main `Reader` API is now a container with `static` and `dynamic` as
sub-readers, as attributes. This allows use of the top-level API methods
which dispatch over the two, or direct use of `dj.read.static` or
`dj.read.dynamic` to bypass the dispatch logic.

This also corrects cases in which the string `"None"` could be produced
by calling `str()` on a result that was nullary.
